Move Share link settings into its own component
