+++ src/app/file-browser/components/sharing-dialog/sharing-dialog.component.ts
@@ -92,47 +92,15 @@ export class SharingDialogComponent implements OnInit {
   public shares: ShareVO[] = [];
   public pendingShares: ShareVO[] = [];
 
-  public shareLink = '';
-
-  public previewToggle: 0 | 1 = 1;
-  public autoApproveToggle: 0 | 1 = 1;
-  public expiration: Expiration;
-  public linkDefaultAccessRole = 'viewer';
-
-  public updatingLink = false;
-  public linkCopied = false;
-  public showLinkSettings = false;
-
-  public linkType = 'private';
-
-  public trashIcon = faTrash;
-
   public newAccessRole: AccessRoleType = 'access.role.viewer';
   public accessRoleOptions: FormInputSelectOption[] =
     SHARE_ACCESS_ROLE_FIELD.selectOptions.reverse();
-  public expirationOptions: FormInputSelectOption[] = EXPIRATION_OPTIONS;
+
   public relationOptions: FormInputSelectOption[] = RELATION_OPTIONS;
 
   public sendingInvitation = false;
   public showInvitationForm = false;
   public invitationForm: UntypedFormGroup;
-
-  public shareLinkTypes = [
-    {
-      text: 'Anyone can view',
-      value: 'public',
-      description: 'Anyone with the link can view and download.',
-    },
-    {
-      text: 'Restricted',
-      value: 'private',
-      description: 'Must be logged in to view.',
-    },
-  ];
-
-  public displayDropdown = false;
-
-  @ViewChild('shareUrlInput', { static: false }) shareUrlInput: ElementRef;
 
   public archiveFilterFn = (a: ArchiveVO) => {
     return (
@@ -149,10 +117,7 @@ export class SharingDialogComponent implements OnInit {
     private api: ApiService,
     private messageService: MessageService,
     private relationshipService: RelationshipService,
-    private ga: GoogleAnalyticsService,
     private route: ActivatedRoute,
-    private feature: FeatureFlagService,
-    private shareLinkService: ShareLinksApiService,
   ) {
     this.invitationForm = this.fb.group({
       fullName: ['', [Validators.required]],
@@ -162,9 +127,6 @@ export class SharingDialogComponent implements OnInit {
     });
 
     this.shareItem = this.data.item as ItemVO;
-    this.shareLinkResponse = this.data.shareLinkResponse;
-
-    this.displayDropdown = feature.isEnabled('unlisted-share');
   }
 
   ngOnInit(): void {
@@ -182,13 +144,6 @@ export class SharingDialogComponent implements OnInit {
     this.canShare = this.shareItem.accessRole === 'access.role.owner';
 
     this.relationshipService.update();
-
-    this.shareLink = shareUrlBuilder(
-      this.data.shareLinkResponse.itemType,
-      this.data.shareLinkResponse.token,
-      this.data.shareLinkResponse.itemId,
-    );
-    this.setShareLinkFormValue();
 
     this.checkQueryParams();
   }
@@ -427,161 +382,6 @@ export class SharingDialogComponent implements OnInit {
           translate: true,
         });
       }
-    }
-  }
-
-  getExpirationFromExpiresDT(expiresDT: string): Expiration {
-    if (!expiresDT) {
-      return Expiration.Never;
-    }
-
-    const diff = differenceInHours(
-      new Date(expiresDT),
-      new Date(this.shareLinkResponse.createdDT),
-    );
-
-    if (diff <= 24 * ExpirationDays.Day) {
-      return Expiration.Day;
-    } else if (diff <= 24 * ExpirationDays.Week) {
-      return Expiration.Week;
-    } else if (diff <= 24 * 7 * ExpirationDays.Month) {
-      return Expiration.Month;
-    } else if (diff <= 24 * 7 * ExpirationDays.Year) {
-      return Expiration.Year;
-    } else {
-      return Expiration.Other;
-    }
-  }
-
-  getExpiresDTFromExpiration(expiration: Expiration): string {
-    switch (expiration) {
-      case Expiration.Never:
-        return null;
-      case Expiration.Day:
-        return getSQLDateTime(
-          addDays(new Date(this.shareLinkResponse.createdDT), 1),
-        );
-      case Expiration.Week:
-        return getSQLDateTime(
-          addDays(new Date(this.shareLinkResponse.createdDT), 7),
-        );
-      case Expiration.Month:
-        return getSQLDateTime(
-          addDays(new Date(this.shareLinkResponse.createdDT), 30),
-        );
-      case Expiration.Year:
-        return getSQLDateTime(
-          addDays(new Date(this.shareLinkResponse.createdDT), 365),
-        );
-    }
-  }
-
-  setShareLinkFormValue(): void {
-    if (this.shareLinkResponse) {
-      console.log(this.shareLinkResponse);
-      this.previewToggle = this.shareLinkResponse.previewToggle;
-      this.autoApproveToggle = this.shareLinkResponse.autoApproveToggle || 0;
-      this.expiration = this.getExpirationFromExpiresDT(
-        this.shareLinkResponse.expiresDT,
-      );
-      console.log(this.shareLinkResponse);
-      console.log(this.shareLinkResponse.permissionsLevel);
-      this.linkDefaultAccessRole = this.shareLinkResponse.permissionLevel;
-      this.expirationOptions = EXPIRATION_OPTIONS.filter((expiration) => {
-        switch (expiration.value) {
-          case Expiration.Never:
-          case Expiration.Other:
-            return true;
-          default:
-            return !isPast(
-              new Date(
-                this.getExpiresDTFromExpiration(expiration.value as Expiration),
-              ),
-            );
-        }
-      });
-    } else {
-      this.previewToggle = 1;
-      this.autoApproveToggle = 1;
-      this.expiration = Expiration.Never;
-      this.expirationOptions = EXPIRATION_OPTIONS;
-    }
-  }
-
-  async generateShareLink() {
-    this.updatingLink = true;
-    try {
-      const response = await this.api.share.generateShareLink(this.shareItem);
-      this.shareLinkResponse = response;
-      this.shareLinkResponse.autoApproveToggle = this.autoApproveToggle || 0;
-      this.shareLinkResponse.previewToggle = this.previewToggle || 0;
-      await this.api.share.updateShareLink(this.shareLinkResponse);
-      this.setShareLinkFormValue();
-      this.showLinkSettings = true;
-      this.ga.sendEvent(EVENTS.SHARE.ShareByUrl.initiated.params);
-    } catch (err) {
-      if (err instanceof ShareResponse) {
-        this.messageService.showError({
-          message: err.getMessage(),
-          translate: true,
-        });
-      }
-    } finally {
-      this.updatingLink = false;
-    }
-  }
-
-  copyShareLink() {
-    const element = this.shareUrlInput.nativeElement as HTMLInputElement;
-
-    copyFromInputElement(element);
-
-    this.linkCopied = true;
-    setTimeout(() => {
-      this.linkCopied = false;
-    }, 5000);
-  }
-
-  async removeShareLink() {
-    const deferred = new Deferred();
-    try {
-      await this.promptService.confirm(
-        'Remove link',
-        'Are you sure you want to remove this link?',
-        deferred.promise,
-        'btn-danger',
-      );
-
-      await this.api.share.removeShareLink(this.shareLinkResponse);
-      this.shareLink = null;
-      this.setShareLinkFormValue();
-      deferred.resolve();
-      this.showLinkSettings = false;
-    } catch (response) {
-      deferred.resolve();
-      if (response instanceof ShareResponse) {
-        this.messageService.showError({ message: response.getMessage() });
-      }
-    }
-  }
-
-  async onShareLinkPropChange(propName: ShareByUrlProps, value: any) {
-    this.updatingLink = true;
-    try {
-      const update = new ShareByUrlVO(this.shareLink);
-      update[propName] = value;
-      await this.api.share.updateShareLink(update);
-      this.shareLink[propName] = update[propName];
-    } catch (err) {
-      if (err instanceof ShareResponse) {
-        this.messageService.showError({
-          message: err.getMessage(),
-          translate: true,
-        });
-      }
-      this.setShareLinkFormValue();
-    } finally {
-      this.updatingLink = false;
     }
   }
 }
